Move comments view classes to individual files
